Limit translationaids results to title matches
translationaids can return other messages with the same source text (if
I understand it correctly). Use only those that come from the same key.
(Arguably we’re abusing the API for something it’s not indended for, but
I didn’t find a better way to get a translate URL or just the message
group.)