Replace language in URL using urllib.parse
This is more robust than the str.replace, and probably the “right” way
to do it. That said, I’m not happy with how it turns : into %3A, and I
haven’t found a way to disable that, so I think I’ll replace this with
another implementation. I’m mainly committing this so I have it in the
history, as backup or whatever.