Use translationaids groups prop
This API was introduced in Gerrit change I4ca7139b11 [1] (already merged
by the time of commit 5de2d28f2d, but not yet deployed then), and ought
to provide us with the most robust way to construct a translate URL:
directly get the groups of the message and build the URL from that
(assuming that the first group is the one we want – for the messages
I’ve tried, the groups seem to be ordered roughly from more to less
specific). The other methods to get a URL are kept for now, but I
suspect they can be removed soon – we emit a warning before we fall back
to them, so that we can see if they’re needed or not in the server log.

[1]: https://gerrit.wikimedia.org/r/c/mediawiki/extensions/Translate/+/692021