Fix message key case for MediaWiki messages
If the user entered the message key in the wrong case (e.g. because they
copied it from a MediaWiki: title), we’ll still be able to find the
message (because intitle: is case-insensitive), but the link we generate
will be broken (because it contains the key with the broken case). For
MediaWiki messages, we can fix this by deriving the key back from the
title, converting the first character to lowercase and leaving the rest
as it is (in the title). For non-MediaWiki messages, that doesn’t feel
like a safe conversion (I have no idea how other i18n systems treat case
in message keys, and if they also covnentionally begin their keys in
lowercase or not), so limit this to MediaWiki messages.